backend/pyproject.toml
[project]
name = "agent-backend"
version = "0.1.0"
description = "AI Agent MVP Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "sse-starlette>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
